Inhabilitació de docker si el so es diferent de Linux #3

+++ bin/docker_cleanup.sh
@@ -10,14 +10,26 @@ set -o errexit
 ###         Cleanup             ###
 ###################################
 
-DOCKER=$(which docker)
-echo "docker at $DOCKER"
-
 echo ""
 PROJECT_PATH="$( cd "$( dirname "${BASH_SOURCE[0]}" )" && cd .. && pwd )"
 echo "Project path at $PROJECT_PATH"
 echo ""
 echo "[$(date +"%Y-%m-%d %T")] Stopping and removing all the containers..."
 echo ""
+
+#source $PROJECT_PATH/bin/lib_string_utils.sh 
+source $PROJECT_PATH/bin/lib_env_utils.sh
+
+#lib_env_utils.loadenv ${PROJECT_PATH}
+#echo ""
+lib_env_utils.check_os
+echo ""
+lib_env_utils.check_docker
+echo ""
+
+if [[ "${DOCKER}" == "/dev/null" ]]; then
+  echo "Docker not installed. Exiting"
+  exit 1
+fi
 
 ${DOCKER} stop $(${DOCKER} ps -a -q) && ${DOCKER} rm $(${DOCKER} ps -a -q) --volumes
