+++ bin/docker_install.sh
@@ -11,15 +11,41 @@ set -o errexit
 ###   DOCKER INSTALL UTILS      ###
 ###################################
 
-sudo apt-get update
-sudo apt-get upgrade
-sudo apt-get remove docker docker-engine docker.io docker-compose
-sudo apt-get autoremove
-sudo apt-get install docker.io docker-compose zip
-sudo systemctl enable docker
-sudo systemctl start docker
-sudo systemctl status docker
-sudo useradd -p $(openssl passwd -1 docker) docker -g docker
-# sudo usermod -a -G docker emiserv
-sudo usermod -a -G docker ${CUSTOM_USERNAME}
-sudo mkdir -p /app/docker
+echo ""
+PROJECT_PATH="$( cd "$( dirname "${BASH_SOURCE[0]}" )" && cd .. && pwd )"
+echo "Project path at $PROJECT_PATH"
+echo ""
+echo "[$(date +"%Y-%m-%d %T")] Installing docker..."
+echo ""
+
+# Taking values from .env file
+
+source $PROJECT_PATH/bin/lib_string_utils.sh 
+source $PROJECT_PATH/bin/lib_env_utils.sh
+
+lib_env_utils.loadenv ${PROJECT_PATH}
+echo ""
+lib_env_utils.check_os
+echo ""
+
+if [[ isLinux -eq 1 ]]; then
+    sudo apt-get update
+    sudo apt-get upgrade
+    sudo apt-get remove docker docker-engine docker.io docker-compose
+    sudo apt-get autoremove
+    sudo apt-get install docker.io docker-compose zip
+    sudo systemctl enable docker
+    sudo systemctl start docker
+    sudo systemctl status docker
+    sudo useradd -p $(openssl passwd -1 docker) docker -g docker
+    # sudo usermod -a -G docker emiserv
+    sudo usermod -a -G docker ${CUSTOM_USERNAME}
+    sudo mkdir -p /app/docker
+else
+    echo ""
+    echo "Docker unavailable"
+    echo ""
+fi
+
+
+
